Tweak  redis docker scritps
Build argument is now ignored since there is not a local Dockerfile to
create the image, it will be always pulled from registry.

Also, add to `docker run` command the `redis` and `redis_db` network
alias to the redis container so as to make redis connections be
configurable via host names instead of human-unfriendly ip addresses.